+++ ios/Classes/Banner/BannerAdController.swift
@@ -5,6 +5,8 @@ class BannerAdController: NSObject {
 
     var bannerView: GADBannerView!
 
+//    var loadRequested: ((MethodChannel.Result) -> Unit)? = null
+
     let id: String
     let channel: FlutterMethodChannel