+++ lib/src/app_open/app_open.dart
@@ -198,9 +198,10 @@ class AppOpenAd extends LoadShowAd<FullScreenAdEvent> {
     }).timeout(
       timeout ?? this.loadTimeout ?? kDefaultLoadTimeout,
       onTimeout: () {
-        onEventController.add({
-          FullScreenAdEvent.loadFailed: AdError.timeoutError,
-        });
+        if (!onEventController.isClosed)
+          onEventController.add({
+            FullScreenAdEvent.loadFailed: AdError.timeoutError,
+          });
         return false;
       },
     );