+++ lib/src/native/layout_builder/linear_layout.dart
@@ -17,6 +17,8 @@ class AdLinearLayout extends AdView {
     AdDecoration decoration,
     double width,
     double height,
+    double elevation,
+    Color elevationColor,
     this.gravity,
   })  : assert(orientation != null),
         super(
@@ -27,6 +29,8 @@ class AdLinearLayout extends AdView {
           decoration: decoration,
           width: width ?? MATCH_PARENT,
           height: height ?? WRAP_CONTENT,
+          elevation: elevation,
+          elevationColor: elevationColor,
         );
 
   Map<String, dynamic> toJson() {